feat: add  on

+++ src/useCases/sensorUseCases/DeleteSensor/deleteSensorUseCase.ts
@@ -1,3 +1,4 @@
+import { errorHelper, ErrorREST } from '@/domain/errors/errorRest'
 import { ISensorsRepository } from '@/infra/repositories/contracts/ISensorsRepository'
 import { IDeleteSensorDTO } from './IDeleteSensorDTO'
 
@@ -7,9 +8,9 @@ export class DeleteSensorUseCase{
     ){}
 
     async execute(data: IDeleteSensorDTO){
-        // const sensorAlreadyExist = await this.sensorsRepository.findByDescription(data.description)
+        const sensorAlreadyExist = await this.sensorsRepository.findByDescription(data.description)
     
-        // if(!sensorAlreadyExist)  throw new Error
+        if(!sensorAlreadyExist)  throw new ErrorREST(errorHelper.notFound)
 
         await this.sensorsRepository.deleteSensor(data.description)
     }
